+++ src/api/news/index.ts
@@ -1,0 +1,3 @@
+export * from "./types";
+export * from "./search";
+export * from "./detail";
